gstreamer: implement pipewire integration using pwaudiosrc/pwaudiosink
The code path is wrapped in #ifdef USE_PIPEWIRE so that it can be toggled
easily at compile time.

The device string is abused to hold the role name, just like it was
with 4A. In the future this may need to be reconsidered. In theory,
we could detect the backend or make it configurable from the upper layer
if we knew exactly what kind of data is in that device string
(is it an ALSA device name? a pipewire role? a pipewire node id?
a pulseaudio device name? ...)

Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>